+++ doc/changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+# 1.2.6 - 2017-08-14
+### Fixed
+* Nginx compatibility issue
+### Changed
+* no more wp-config.php symlink, included in /web
+
 # 1.2.5 - 2017-08-08
 ### Added
 * Rocket Term model